Add sketch uses MPU to turn on light
Adding a sketch that lights one of the LEDs when the XAxis of the
accelerometer is between a given range. The idea here is that
we might only want to turn on the OLED  when we are looking at
it so we can use the X axis on the accel to give us an idea
of when the watch is in standard watch reading position.