Change up time calculation to use RTC
The up time calculation I was using originally used the millis
function however the millis function stops counting when the
processor goes to sleep so I can't use that to get an accurate
up time. So I added a new block of code that will used the
RTC time instead. I kept the block that uses millis and changed
it to note it was the amount of awake time.